Refactor DeferralForm for more general use
