+++ ui/apps/platform/src/Containers/Vulnerabilities/WorkloadCves/components/ExceptionRequestModal/ExceptionRequestForm.tsx
@@ -1,41 +1,23 @@
 import React, { useState } from 'react';
 import {
-    Bullseye,
     Button,
-    DatePicker,
     Flex,
     FormGroup,
-    FormHelperText,
     Form,
-    HelperText,
-    HelperTextItem,
-    Radio,
-    Spinner,
     Tabs,
     Tab,
     TextArea,
     Text,
     TabContent,
 } from '@patternfly/react-core';
-import { ExclamationCircleIcon } from '@patternfly/react-icons';
 import { FormikHelpers, useFormik } from 'formik';
-import { addDays } from 'date-fns';
 
-import EmptyStateTemplate from 'Components/PatternFly/EmptyStateTemplate';
-import { fetchVulnerabilitiesExceptionConfig } from 'services/ExceptionConfigService';
-import useRestQuery from 'hooks/useRestQuery';
-import { getAxiosErrorMessage } from 'utils/responseErrorUtils';
-
-import {
-    DeferralValues,
-    ScopeContext,
-    deferralValidationSchema,
-    futureDateValidator,
-} from './utils';
+import { ScopeContext, exceptionValidationSchema, ExceptionValues } from './utils';
 import ExceptionScopeField, { ALL } from './ExceptionScopeField';
 import CveSelections, { CveSelectionsProps } from './CveSelections';
+import ExpiryField from './ExpiryField';
 
-function getDefaultValues(cves: string[], scopeContext: ScopeContext): DeferralValues {
+function getDefaultValues(cves: string[], scopeContext: ScopeContext): ExceptionValues {
     const imageScope =
         scopeContext === 'GLOBAL'
             ? { registry: ALL, remote: ALL, tag: ALL }
@@ -43,21 +25,14 @@ function getDefaultValues(cves: string[], scopeContext: ScopeContext): DeferralV
 
     return { cves, comment: '', scope: { imageScope } };
 }
-
-/**
- * Returns the date portion of an ISO date string
- * @param date - ISO date string
- * @returns Date portion of the ISO date string, or an empty string if the date is falsy
- */
-function prettifyDate(date = ''): string {
-    return date.substring(0, 10);
-}
-
 export type ExceptionRequestFormProps = {
     cves: CveSelectionsProps['cves'];
     scopeContext: ScopeContext;
-    onSubmit: (formValues: DeferralValues, helpers: FormikHelpers<DeferralValues>) => void;
+    onSubmit: (formValues: ExceptionValues, helpers: FormikHelpers<ExceptionValues>) => void;
     onCancel: () => void;
+    formHeaderText: string;
+    commentFieldLabel: string;
+    showExpiryField?: boolean;
 };
 
 function ExceptionRequestForm({
@@ -65,9 +40,11 @@ function ExceptionRequestForm({
     scopeContext,
     onSubmit,
     onCancel,
+    formHeaderText,
+    commentFieldLabel,
+    showExpiryField = false,
 }: ExceptionRequestFormProps) {
     const [activeKeyTab, setActiveKeyTab] = useState<string | number>('options');
-    const { data: config, loading, error } = useRestQuery(fetchVulnerabilitiesExceptionConfig);
 
     const formik = useFormik({
         initialValues: getDefaultValues(
@@ -75,46 +52,11 @@ function ExceptionRequestForm({
             scopeContext
         ),
         onSubmit,
-        validationSchema: deferralValidationSchema,
+        validationSchema: exceptionValidationSchema,
     });
-    const {
-        values,
-        setValues,
-        handleBlur,
-        setFieldValue,
-        touched,
-        handleSubmit,
-        submitForm,
-        isSubmitting,
-        errors,
-    } = formik;
 
-    if (loading) {
-        return (
-            <Bullseye>
-                <Spinner isSVG />
-            </Bullseye>
-        );
-    }
-
-    if (error) {
-        return (
-            <Bullseye>
-                <EmptyStateTemplate
-                    headingLevel="h2"
-                    title="There was an error loading the vulnerability exception configuration"
-                    icon={ExclamationCircleIcon}
-                    iconClassName="pf-u-danger-color-100"
-                >
-                    {getAxiosErrorMessage(error)}
-                </EmptyStateTemplate>
-            </Bullseye>
-        );
-    }
-
-    function setExpiry(expiry: DeferralValues['expiry']) {
-        return setValues((prev) => ({ ...prev, expiry }));
-    }
+    const { handleBlur, setFieldValue, touched, handleSubmit, submitForm, isSubmitting, errors } =
+        formik;
 
     return (
         <>
@@ -140,116 +82,15 @@ function ExceptionRequestForm({
                         direction={{ default: 'column' }}
                         spaceItems={{ default: 'spaceItemsLg' }}
                     >
-                        <Text>CVEs will be marked as deferred after approval</Text>
-                        {config && (
-                            <FormGroup label="How long should the CVEs be deferred?" isRequired>
-                                <Flex
-                                    direction={{ default: 'column' }}
-                                    spaceItems={{ default: 'spaceItemsXs' }}
-                                >
-                                    {config.expiryOptions.fixableCveOptions.anyFixable && (
-                                        <Radio
-                                            id="any-cve-fixable"
-                                            name="any-cve-fixable"
-                                            isChecked={values.expiry?.type === 'ANY_CVE_FIXABLE'}
-                                            onChange={() => setExpiry({ type: 'ANY_CVE_FIXABLE' })}
-                                            label="When any CVE is fixable"
-                                        />
-                                    )}
-                                    {config.expiryOptions.fixableCveOptions.allFixable && (
-                                        <Radio
-                                            id="all-cve-fixable"
-                                            name="all-cve-fixable"
-                                            isChecked={values.expiry?.type === 'ALL_CVE_FIXABLE'}
-                                            onChange={() => setExpiry({ type: 'ALL_CVE_FIXABLE' })}
-                                            label="When all CVEs are fixable"
-                                        />
-                                    )}
-                                    {config.expiryOptions.dayOptions
-                                        .filter((option) => option.enabled)
-                                        .map(({ numDays }) => (
-                                            <Radio
-                                                id={`fixed-duration-${numDays}`}
-                                                name={`fixed-duration-${numDays}`}
-                                                key={`fixed-duration-${numDays}`}
-                                                isChecked={
-                                                    values.expiry?.type === 'TIME' &&
-                                                    values.expiry?.days === numDays
-                                                }
-                                                onChange={() =>
-                                                    setExpiry({
-                                                        type: 'TIME',
-                                                        days: numDays,
-                                                    })
-                                                }
-                                                label={`For ${numDays} days`}
-                                            />
-                                        ))}
-                                    {/* TODO - Awaiting backend support for indefinite deferrals
-                                         config.expiryOptions.indefinite && (
-                                            <Radio
-                                                id="indefinite"
-                                                name="indefinite"
-                                                isChecked={values.expiry?.type === 'INDEFINITE'}
-                                                onChange={() => {}}
-                                                label="Indefinitely"
-                                            />
-                                        )
-                                        */}
-                                    {config.expiryOptions.customDate && (
-                                        <Radio
-                                            id="custom-date"
-                                            name="custom-date"
-                                            isChecked={values.expiry?.type === 'CUSTOM_DATE'}
-                                            onChange={() =>
-                                                setExpiry({
-                                                    type: 'CUSTOM_DATE',
-                                                    date: addDays(new Date(), 1).toISOString(),
-                                                })
-                                            }
-                                            label="Until a specific date"
-                                        />
-                                    )}
-
-                                    {config.expiryOptions.customDate &&
-                                        values.expiry?.type === 'CUSTOM_DATE' && (
-                                            <div>
-                                                <DatePicker
-                                                    name="custom-date-picker"
-                                                    value={prettifyDate(values.expiry.date)}
-                                                    onChange={(_, value) =>
-                                                        setExpiry({
-                                                            type: 'CUSTOM_DATE',
-                                                            date: new Date(value).toISOString(),
-                                                        })
-                                                    }
-                                                    validators={[futureDateValidator]}
-                                                />
-                                            </div>
-                                        )}
-                                    {errors.expiry && (
-                                        <FormHelperText isError isHidden={false}>
-                                            <HelperText>
-                                                <HelperTextItem
-                                                    variant="error"
-                                                    icon={<ExclamationCircleIcon />}
-                                                    className="pf-u-display-flex pf-u-align-items-center"
-                                                >
-                                                    {errors.expiry}
-                                                </HelperTextItem>
-                                            </HelperText>
-                                        </FormHelperText>
-                                    )}
-                                </Flex>
-                            </FormGroup>
-                        )}
+                        <Text>{formHeaderText}</Text>
+                        {showExpiryField && <ExpiryField formik={formik} />}
                         <ExceptionScopeField
                             fieldId="scope"
                             label="Scope"
                             formik={formik}
                             scopeContext={scopeContext}
                         />
-                        <FormGroup fieldId="comment" label="Deferral rationale" isRequired>
+                        <FormGroup fieldId="comment" label={commentFieldLabel} isRequired>
                             <TextArea
                                 id="comment"
                                 name="comment"
